@@ -13,11 +13,8 @@ use Symfony\Component\Form\DataTransformerInterface;
 
 class NodesTagsTransformer implements DataTransformerInterface
 {
-    private ManagerRegistry $managerRegistry;
-
-    public function __construct(ManagerRegistry $managerRegistry)
+    public function __construct(private readonly ManagerRegistry $managerRegistry)
     {
-        $this->managerRegistry = $managerRegistry;
     }
 
     /**
@@ -25,6 +22,7 @@ class NodesTagsTransformer implements DataTransformerInterface
      *
      * @return int[]
      */
+    #[\Override]
     public function transform(mixed $value): array
     {
         $ids = [];
@@ -42,6 +40,7 @@ class NodesTagsTransformer implements DataTransformerInterface
      *
      * @return Collection<int, NodesTags>
      */
+    #[\Override]
     public function reverseTransform(mixed $value): Collection
     {
         $nodesTags = [];
